+++ App.js
@@ -1,17 +1,12 @@
 import React, { useState, useEffect } from 'react';
-import { ScrollView } from 'react-native';
-
+import { ScrollView, View } from 'react-native';
 import AuthScreen from './screens/auth/AuthScreen/authScr.js';
+import ProfileScreen from './screens/profile/profileScr.js';
 import AuthenticatedScreen from './screens/auth/AuthenticatedScreen/authenticatedScr.js';
 import SplashScreen from './screens/auth/AuthScreen/splashScr.js';
-import ProfileScreen from './screens/profile/screens/profileScr.js';
-import NotebookScreen from './screens/notebook/notebookScr.js';
-import HomeScreen from './screens/home/screens/homeScr.js';
 import LibraryScreen from './screens/library/libraryScr.js';
-import ForumScreen from './screens/forum/forumScr.js';
-
-import { onAuthStateChanged, signOut } from 'firebase/auth';
-import { auth } from './initializeFB.js'; // Use the configured auth
+import { createUserWithEmailAndPassword, signInWithEmailAndPassword, onAuthStateChanged, signOut } from 'firebase/auth';
+import { auth } from './initializeFB'; // Use the configured auth
 import { styles } from './components/styles.js';
 
 const App = () => {
@@ -40,6 +35,46 @@ const App = () => {
     resetInputFields();
   }, [isLogin]);
 
+  const handleAuthentication = async () => {
+    try {
+      if (user) {
+        // If user is already authenticated, log out
+        console.log('You have logged out successfully!');
+        await signOut(auth);
+        setCurrentScreen('Auth');
+      } else {
+        // Log in or sign up
+        if (isLogin) {
+          // Log in
+          await signInWithEmailAndPassword(auth, email, password);
+          console.log('You have signed in successfully!');
+          setCurrentScreen('Authenticated');
+        } else {
+          // Sign up
+          if (password !== retypePassword) {
+            console.error("Passwords don't match");
+            return;
+          }
+          await createUserWithEmailAndPassword(auth, email, password);
+          console.log('You have created an account successfully!');
+          setCurrentScreen('Authenticated');
+        }
+      }
+    } catch (error) {
+      console.error('Authentication error:', error.message);
+    }
+  };
+
+  const handleSignOut = async () => {
+    try {
+      await signOut(auth);
+      console.log('You have signed out successfully!');
+      setCurrentScreen('Auth');
+    } catch (error) {
+      console.error('Sign out error:', error.message);
+    }
+  }
+
   const resetInputFields = () => {
     setUsername('');
     setEmail('');
@@ -55,35 +90,12 @@ const App = () => {
     return <SplashScreen onTimeout={handleSplashScreenTimeout} />;
   }
 
-  const handleSignOut = async () => {
-    try {
-      await signOut(auth);
-      console.log('You have signed out successfully!');
-      setCurrentScreen('Auth');
-      setIsLogin(true);
-    } catch (error) {
-      console.error('Sign out error:', error.message);
-    }
-  };
-
   const directToProfile = () => {
     setCurrentScreen('Profile');
   };
 
-  const directToNotebook = () => {
-    setCurrentScreen('Notebook')
-  };
-
-  const directToHome = () => {
-    setCurrentScreen('Home');
-  };
-
   const directToLibrary = () => {
     setCurrentScreen('Library');
-  };
-
-  const directToForum = () => {
-    setCurrentScreen('Forum');
   };
 
   return (
@@ -100,56 +112,26 @@ const App = () => {
           setRetypePassword={setRetypePassword}
           isLogin={isLogin}
           setIsLogin={setIsLogin}
-          setCurrentScreen={setCurrentScreen}
+          handleAuthentication={handleAuthentication}
         />
       )}
       {currentScreen === 'Authenticated' && (
         <AuthenticatedScreen
           user={user}
-          handleAuthentication={() => setCurrentScreen('Auth')}
-          directToHome={directToHome}
+          handleAuthentication={handleAuthentication}
+          navigateToProfile={() => setCurrentScreen('Profile')}
         />
       )}
       {currentScreen === 'Profile' && (
         <ProfileScreen
           user={user}
           handleSignOut={handleSignOut}
-          directToNotebook={directToNotebook}
-          directToHome={directToHome}
           directToLibrary={directToLibrary}
-          directToForum={directToForum}
-        />
-      )}
-      {currentScreen === 'Notebook' && (
-        <NotebookScreen
-          directToProfile={directToProfile}
-          directToHome={directToHome}
-          directToLibrary={directToLibrary}
-          directToForum={directToForum}
-        />
-      )}
-      {currentScreen === 'Home' && (
-        <HomeScreen
-          directToProfile={directToProfile}
-          directToNotebook={directToNotebook}
-          directToLibrary={directToLibrary}
-          directToForum={directToForum}
         />
       )}
       {currentScreen === 'Library' && (
         <LibraryScreen
           directToProfile={directToProfile}
-          directToNotebook={directToNotebook}
-          directToHome={directToHome}
-          directToForum={directToForum}
-        />
-      )}
-      {currentScreen === 'Forum' && (
-        <ForumScreen
-          directToProfile={directToProfile}
-          directToNotebook={directToNotebook}
-          directToHome={directToHome}
-          directToLibrary={directToLibrary}
         />
       )}
     </ScrollView>
