+++ package.json
@@ -25,11 +25,8 @@
     "react": "18.2.0",
     "react-native": "0.74.1",
     "react-native-dotenv": "^3.4.11",
-    "react-native-gesture-handler": "^2.17.1",
-    "react-native-modern-datepicker": "^1.0.0-beta.91",
-    "react-native-reanimated": "^3.12.1",
-    "react-native-safe-area-context": "^4.10.1",
-    "react-native-swiper": "^1.6.0"
+    "react-native-gesture-handler": "^2.16.2",
+    "react-native-safe-area-context": "^4.10.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
