+++ screens/auth/AuthScreen/authScr.js
@@ -1,13 +1,7 @@
 import React from 'react';
-import { View, Text, Image, TouchableOpacity, StyleSheet, Dimensions } from 'react-native';
-import { signInWithEmailAndPassword, createUserWithEmailAndPassword } from 'firebase/auth';
-
-import { auth, db } from '../../../initializeFB';
-import { doc, setDoc, addDoc, collection } from 'firebase/firestore';
-
 import LogInScreen from './logIn';
 import SignUpScreen from './signUp';
-
+import { View, Text, Image, TouchableOpacity, StyleSheet, Dimensions } from 'react-native';
 
 const AuthScreen = ({
   username, setUsername,
@@ -15,51 +9,11 @@ const AuthScreen = ({
   password, setPassword,
   retypePassword, setRetypePassword,
   isLogin, setIsLogin,
-  setCurrentScreen
+  handleAuthentication
 }) => {
-  const { height } = Dimensions.get('window');
+  const { width, height } = Dimensions.get('window');
   // image
-  const imageSize = height * 0.2;
-
-  const handleAuthentication = async () => {
-    try {
-      if (isLogin) {
-        await signInWithEmailAndPassword(auth, email, password);
-        setCurrentScreen('Authenticated');
-        console.log('You have signed in successfully!');
-      } else {
-        if (password !== retypePassword) {
-          console.error("Passwords don't match");
-          return;
-        }
-        const userCredential = await createUserWithEmailAndPassword(auth, email, password);
-        const user = userCredential.user;
-
-        await setDoc(doc(db, 'users', user.uid), {
-          username: username,
-          email: email,
-          picture: require('../../../assets/profile_images/default_profile_picture.png'),
-          description: ''
-        });
-
-        const petsCollectionRef = collection(db, 'users', user.uid, 'pets');
-        await addDoc(petsCollectionRef, {
-          name: '',
-          picture: require('../../../assets/home_images/default_pet_image_square.png'),
-          breed: '',
-          birthDate: '',
-          age: '',
-          gender: '',
-          notes: '',
-        });
-
-        setCurrentScreen('Authenticated');
-        console.log('You have created an account successfully!');
-      }
-    } catch (error) {
-      console.error('Authentication error:', error.message);
-    }
-  };
+  const imageSize = height * 0.2; // size of image (dogs and cats)
 
   return (
     <View style={styles.authContainer}>
