+++ screens/auth/AuthenticatedScreen/authenticatedScr.js
@@ -1,54 +1,35 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { View, Text, StyleSheet, Image, Dimensions } from 'react-native';
 
-import { db } from '../../../initializeFB';
-import { doc, getDoc } from 'firebase/firestore';
-
-const AuthenticatedScreen = ({ user, directToHome }) => {
-  const [userData, setUserData] = useState(null);
-
+const AuthenticatedScreen = ({ user, navigateToProfile }) => {
   useEffect(() => {
-    const fetchUserData = async () => {
-      try {
-        const docRef = doc(db, 'users', user.uid);
-        const docSnap = await getDoc(docRef);
-        if (docSnap.exists()) {
-          setUserData(docSnap.data());
-        } else {
-          console.log('No such document!');
-        }
-      } catch (error) {
-        console.error('Error fetching user data:', error.message);
-      }
-    };
-
-    fetchUserData();
-
     const timer = setTimeout(() => {
-      directToHome();
-    }, 2000);
+      navigateToProfile();
+    }, 2000); // Redirect after 2 seconds
 
     return () => clearTimeout(timer);
-  }, [user]);
+  }, []);
 
   const { width, height } = Dimensions.get('window');
-  const fontSize = height * 0.055;
-  const marginVertical = fontSize * 0.05;
-  const usernameFontSize = height * 0.025;
-  const imageSize = height * 0.58;
+  // title
+  const fontSize = height * 0.055; // size of title
+  const marginVertical = fontSize * 0.05; // vertical margin of title to be 5% of font size
+  // email text
+  const usernameFontSize = height * 0.025; // size of email text (adjusted for more appropriate size)
+  // image
+  const imageSize = height * 0.58; // size of image (dogs and cats)
   const imagePosition = {
     bottom: 0.01 * imageSize,
     left: -0.15 * width,
-  };
+  }; // position of image
+  // title container
   const titleCmarginTop = height * 0.08;
 
   return (
     <View style={styles.container}>
       <View style={[styles.titleContainer, { marginTop: titleCmarginTop }]}>
         <Text style={[styles.title, { fontSize, lineHeight: fontSize, marginVertical }]}>Welcome</Text>
-        <Text style={[styles.usernameText, { fontSize: usernameFontSize, lineHeight: usernameFontSize, marginVertical }]}>
-          {userData && userData.username}
-        </Text>
+        <Text style={[styles.usernameText, { fontSize: usernameFontSize, lineHeight: usernameFontSize, marginVertical }]}>{user.username}</Text>
       </View>
       <Image
         source={require('../../../assets/app_images/authenticatedScr_Background.png')}
@@ -62,19 +43,19 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: '#FCF9D9',
-    justifyContent: 'flex-start',
-    alignItems: 'flex-start',
+    justifyContent: 'flex-start', // Align content to the top
+    alignItems: 'flex-start', // Align content to the left
     padding: 20,
   },
   titleContainer: {
     marginTop: '3%',
-    marginLeft: 5,
+    marginLeft: 5, // No margin from the left
   },
   title: {
     fontWeight: 'bold',
   },
   usernameText: {
-    marginTop: 5,
+    marginTop: 5, // Space between title and email
   },
   image: {
     position: 'absolute',
